# Improve IOBuffer read performance