@JeffBezanson 2015-07-29
Wow, nice.

Seems like a slightly large function to inline. Maybe the other optimizations are sufficient?