@jiahao 2015-07-29
:+1: 

This performance issue was discovered by profiling JuliaLang/GZip.jl#32, where we found that ~40% of the time of the buffered GZip read was spent in the bowels of iobuffer.jl.