@carnaval 2015-07-29
Given the typical `while !eof; read(UInt8);` use case, it's pretty bad to not inline in those case (reload all the object to do the checks every time).

It's probably a bit wasteful if you manually read a lot of bytes in, say, a parser. One could hope that the inliner could recognize the loop but we're not quite there yet.