@simonster 2015-07-29
Without inlining there is no performance difference here. But OTOH inlining is only worthwhile in loops where it actually allows LLVM to hoist out the loads, which at the moment means that there can't really be any other stores in the loop. So it's possible that this makes a bigger difference in the trivial benchmark above than it would actually make in the real world.