@jiahao 2015-07-29
Another time sink is that I/O functions like `read(::IOBuffer)` checks if the IOBuffer is readable at runtime, and throws an error if it isn't. I doubt that `IOBuffer` attributes like `readability` ordinarily change at runtime, if at all. Nonetheless in the current implementation, `IOBuffer.readability` is a Boolean field whose value has to be checked at runtime, and the cost of which adds up when doing Big Data™ I/O operations. Removing this check seems more difficult though.