@kmsquire 2015-07-29
> Nonetheless in the current implementation, IOBuffer.readability is a Boolean field whose value has to be checked at runtime, and the cost of which adds up when doing Big Data™ I/O operations. 

Perhaps make `IOBuffer` (and/or `IO` in general) parameterized based on read/write?  Do some streams change their readability/writability throughout their life?