@JeffBezanson 2015-07-29
That is a good point; do we really need to support write-only buffers? I recall the student manual to Art of Electronics includes a data sheet for a write-only memory because it is absurd and hilarious.