@ScottPJones 2015-07-29
:+1: Love to see the performance improvements, this will definitely speed up some of our (company I'm consulting for) code!