@vtjnash 2015-07-31
> Seems like a slightly large function to inline. Maybe the other optimizations are sufficient?

it's two bool comparisons, a pointer dereference, and an integer increment. that's pretty small and definitely worth inlining. i'm a bit surprised the inliner heuristics didn't figure that out on its own though.