@vtjnash 2015-07-31
also, fwiw, based on my recollection of initial benchmarking of the IOBuffer implementation, I believe this makes it equivalent in performance to the IOString object.